backend/requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
cairocffi==1.7.1
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
fastapi==0.115.0
filelock==3.20.0
fonttools==4.60.1
freetype-py==2.5.1
fsspec==2025.10.0
h11==0.16.0
hydra-core==1.3.2
idna==3.11
imageio==2.37.0
intel-openmp==2021.4.0
iopath==0.1.10
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
mkl==2021.4.0
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.8.1.78
packaging==25.0
pillow==12.0.0
portalocker==3.2.0
pycairo==1.29.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.3
rlPyCairo==0.4.0
SAM-2 @ git+https://github.com/facebookresearch/sam2.git@2b90b9f5ceec907a1c18123530e92e794ad901a4
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.38.6
sympy==1.13.1
tbb==2021.13.1
threadpoolctl==3.6.0
tifffile==2025.10.16
tinycss2==1.4.0
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.30.0
webencodings==0.5.1
wheel==0.45.1
